requirements.txt
# Core dependencies
pyyaml>=6.0
boto3>=1.26.0
mysql-mimic>=2.5.0

# Testing dependencies
pytest>=7.0.0
hypothesis>=6.0.0
pytest-cov>=4.0.0
moto>=4.0.0
